RF_01.001 | New Item | NewItemPage2Test - refactoring NewItemPage2Test was deleted
