+++ src/test/java/school/redrover/FolderTest.java
@@ -4,6 +4,7 @@ import org.testng.Assert;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
+import school.redrover.common.TestUtils;
 import school.redrover.page.HomePage;
 import school.redrover.page.folder.FolderProjectPage;
 import school.redrover.testdata.TestDataProvider;
@@ -13,6 +14,7 @@ import java.util.List;
 public class FolderTest extends BaseTest {
 
     private static final String FOLDER_NAME = "ProjectFolder";
+    private static final String FOLDER_SECOND_NAME = "ProjectFolder2";
     private static final String FOLDER_DISPLAY_NAME = "Folder Display Name";
     private static final String ITEM_NAME = "Item1";
 
@@ -72,6 +74,24 @@ public class FolderTest extends BaseTest {
 
         Assert.assertEquals(folderProjectPage.getProjectNameList().get(0), ITEM_NAME);
         Assert.assertEquals(folderProjectPage.getProjectNameList().size(), 1);
+    }
+
+    @Test(dependsOnMethods = "testCreateFreestyleProjectInFolder")
+    public void testIfTwoDifferentFoldersCanHoldItemsWithTheSameNames() {
+
+        String folderProjectName = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(FOLDER_SECOND_NAME)
+                .selectFolderAndClickOk()
+                .getHeader()
+                .clickLogoIcon()
+                .clickOnNewItemLinkWithChevron(FOLDER_SECOND_NAME)
+                .sendItemName(ITEM_NAME)
+                .selectFolderAndClickOk()
+                .clickSave()
+                .getProjectName();
+
+        Assert.assertEquals(folderProjectName, ITEM_NAME);
     }
 
     @Test
