+++ src/test/java/school/redrover/NewItemTest.java
@@ -2,16 +2,20 @@ package school.redrover;
 
 
 import org.testng.Assert;
+import school.redrover.common.TestUtils;
 import school.redrover.page.HomePage;
 import school.redrover.page.newitem.NewItemPage;
 import school.redrover.testdata.TestDataProvider;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 
+import java.util.List;
+
 
 public class NewItemTest extends BaseTest {
 
     private static final String ITEM_NAME = "ItemName";
+    private static final String ITEM_NAME_NEXT = "ItemName2";
     private static final String RED_COLOR_ERROR = "rgba(230, 0, 31, 1)";
 
     @Test
@@ -84,5 +88,79 @@ public class NewItemTest extends BaseTest {
 
         Assert.assertEquals(newItemPage.getItemNameInvalidMessage(), "» A job already exists with the name ‘%s’".formatted(ITEM_NAME));
         Assert.assertEquals(newItemPage.getItemNameInvalidMessageColor(), RED_COLOR_ERROR);
+    }
+
+    @Test(dataProvider = "itemDescriptions", dataProviderClass = TestDataProvider.class)
+    public void testItemsDescriptions(List<String> expectedItemDescriptions) {
+        List<String> actualItemDescriptions = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .getJobsDescriptions();
+
+        Assert.assertEquals(actualItemDescriptions, expectedItemDescriptions);
+    }
+
+    @Test(dataProvider = "itemTypes", dataProviderClass = TestDataProvider.class)
+    public void testIfSelectedItemIsHighlighted(String itemTypeName) {
+        NewItemPage newItemPage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .clickOnJobItem(itemTypeName);
+
+        Assert.assertTrue(newItemPage.isListItemHighlighted(itemTypeName));
+    }
+
+    @Test
+    public void testIfCopyFromOptionIsDisplayed() {
+        NewItemPage newItemPage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(ITEM_NAME)
+                .selectFolderAndClickOk()
+                .getHeader()
+                .clickLogoIcon()
+                .clickNewItemOnLeftSidePanel();
+
+        Assert.assertEquals(newItemPage.getCopyFromFieldText(), "Copy from");
+        Assert.assertTrue(newItemPage.isCopyFromOptionInputDisplayed());
+    }
+
+    @Test
+    public void testIfCopyFromOptionInputIsNotDisplayed() {
+        NewItemPage newItemPage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel();
+
+        Assert.assertFalse(newItemPage.isCopyFromOptionInputDisplayed());
+    }
+
+    @Test(dependsOnMethods = "testIfCopyFromOptionIsDisplayed")
+    public void testAutocompleteOption() {
+        String actualProjectName = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(ITEM_NAME)
+                .enterValueToCopyFromInput(ITEM_NAME_NEXT)
+                .getDropdownItemText();
+
+        Assert.assertEquals(actualProjectName, ITEM_NAME_NEXT);
+    }
+
+    @Test(dependsOnMethods = "testIfCopyFromOptionIsDisplayed")
+    public void testIfNoItemsMessageIsDisplayed() {
+        String noItemsMessage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(ITEM_NAME)
+                .enterValueToCopyFromInput(ITEM_NAME_NEXT.concat("a"))
+                .getDropdownItemText();
+
+        Assert.assertEquals(noItemsMessage, "No items");
+    }
+
+    @Test(dependsOnMethods = "testIfCopyFromOptionIsDisplayed")
+    public void testIfUserRedirectedToErrorPage() {
+        String errorPageHeading = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(ITEM_NAME)
+                .enterValueToCopyFromInput(ITEM_NAME_NEXT.concat("a"))
+                .clickOkButtonWithError()
+                .getTitle();
+
+        Assert.assertEquals(errorPageHeading, "Error");
     }
 }
